Proyecto_Randomlocke/requirements.txt
streamlit
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow